+++ assert/CSS/pokemon_details/pokemon_details.css
@@ -45,7 +45,6 @@ div > img {
 }
 
 .pokedex-details-id-of-pokemon {
-  border: 1.5px black solid;
   width: 80px;
   height: 80px;
   margin-left: 20px;
